alea-gsw/apps/ground-station-client/pyproject.toml
[tool.poetry]
name = "ground-station-client"
version = "0.1.0"
description = ""
authors = ["SatNOGS project <dev@satnogs.org>", "Evan Feng <21437106-evanfeng@users.noreply.gitlab.com>"]
license = "AGPLv3"
# readme = "README.md"
packages = [
    { include = "groundstationclient" },
]

[tool.poetry.dependencies]
python = "^3.10"
apscheduler = "~3.9"
python-dotenv = ">=0.20.0,<0.21.0"
requests = "^2.32.3"
validators = ">=0.20.0,<0.21.0"
python-dateutil = ">=2.8.0,<2.9.0"
ephem = ">=4.1.3.0,<4.2.0.0"
pytz = "^2024.1"
numpy = ">=1.23.0,<1.24.0"
h5py = ">=3.11.0,<3.12.0"
matplotlib = ">=3.5.0,<3.6.0"
sentry-sdk = ">=1.6.0,<1.7.0"
flask = ">=3.0.3,<3.1.0"
openapi-core = ">=0.19.2,<0.20.0"
sqlalchemy = ">=1.4.53,<1.5.0"

[tool.poetry.group.dev.dependencies]
yapf = "^0.40.2"
pylint = ">=2.14.0,<2.15.0"
tox = ">=4.5.0,<4.6.0"
isort = ">=5.10.0,<5.11.0"
flake8 = ">=4.0.0,<4.1.0"
flake8-quotes = ">=3.3.0,<3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
use_parentheses = "True"
skip_glob = ["docs",".tox"]
line_length = 99
multi_line_output = 2

[tool.pylint.'MASTER']
load-plugins = ["pylint.extensions.docparams", "pylint.extensions.no_self_use"]

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "consider-using-f-string",
    "useless-object-inheritance",
    "consider-using-from-import",
    "duplicate-code",
    "too-many-instance-attributes",
    "too-many-branches",
    "too-many-arguments",
    "too-many-locals",
    "consider-using-with",
    "arguments-differ", # (due to https://github.com/PyCQA/pylint/issues/1482, fixed in pylint>=2.5.0)
    "deprecated-module",
]

[tool.yapf]
column_limit = 99

[tool.yapfignore]
ignore_patterns = [
    ".tox",
    "build",
    "docs",
]